changed main.jsx
