+++ src/main.jsx
@@ -5,6 +5,6 @@ import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
-        <App2 />
+        <App />
     </React.StrictMode>
 );
